django-celery-api/pyproject.toml
[project]
name = "django-ocr-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adrf==0.1.9",
    "asyncio==3.4.3",
    "django==5.1.5",
    "djangorestframework==3.15.2",
    "drf-spectacular==0.28.0",
    "drf-spectacular-sidecar==2024.12.1",
    "dvc==3.59.0",
    "dvc-s3==3.2.0",
    "gunicorn==23.0.0",
    "onnxruntime==1.20.1",
    "paddleocr==2.9.1",
    "paddlepaddle==3.0.0b2",
    "pillow==11.1.0",
    "pymupdf==1.25.2",
    "ruff==0.9.2",
    "uvicorn==0.34.0",
]
